backend/requirements.txt
# Text-to-SQL RAG Chatbot - Backend Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
pydantic-settings==2.1.0

# Database
pymysql==1.1.0
psycopg2-binary==2.9.9
SQLAlchemy==2.0.25

# Embeddings & Vector DB (FAISS = local, no API key)
openai==1.12.0
faiss-cpu>=1.8.0
sentence-transformers==2.3.1

# LLM (Groq = cloud, no local server)
groq>=0.4.0

# LLM (optional)
langchain==0.1.9
langchain-openai==0.0.6
langchain-community==0.0.24

# RAGAS Evaluation
ragas==0.1.6
datasets==2.17.1

# SQL parsing & validation
sqlparse==0.5.0

# Redis (optional - sync jobs, schema cache, chat cache)
redis>=5.0.0

# Utils
httpx==0.26.0
numpy==1.26.4
